Move AddIssues to shared
