+++ src/applications/appeals/10182/content/addIssue.jsx
@@ -2,9 +2,6 @@ import React from 'react';
 
 import { MAX_LENGTH } from '../../shared/constants';
 
-const hintText =
-  'You can only add an issue that you’ve received a VA decision notice for.';
-
 export const issueErrorMessages = {
   missingIssue: 'You must add an issue',
   maxLength: `You can enter a maximum of ${
@@ -20,31 +17,6 @@ export const issueErrorMessages = {
     'You must add an issue with a decision date that’s less than a year old',
   newerDate:
     'You must add an issue with a decision date that’s less than 100 years old',
-};
-
-export const content = {
-  title: {
-    add: 'Add an issue',
-    edit: 'Edit an issue',
-  },
-
-  button: {
-    cancel: 'Cancel',
-    add: 'Add issue',
-    edit: 'Update issue',
-  },
-  name: {
-    label: 'Name of issue',
-    hintText,
-    hint: (
-      <p className="vads-u-font-weight--normal label-description">{hintText}</p>
-    ),
-  },
-  date: {
-    label: 'Date of decision',
-    hint:
-      'You can find the decision date on your decision notice (the letter you received physically in the mail from us).',
-  },
 };
 
 export const missingIssuesErrorMessageText =
